back-end/requirements.txt
flask
flask-cors==3.0.10
flask-socketio==5.3.3
flask-jwt-extended==4.4.4
pymongo==4.3.3
bcrypt==4.0.1
python-dateutil==2.8.2
gunicorn==20.1.0
eventlet==0.33.3
Pillow